Add files via upload
This is a script that will ask if you want to scan a url or ip. 
It will then prompt you for the ip or url that you want scanned. 
You will have to get your own API key from VirusTotal to use it. 
I saved my API key in another file in the same directory and then imported it.