Dice.py and die_cli.py
The first prompts the user for a number of rolls and number of sides on the die to be rolled. It uses a function to test if the input is a number. 
die_cli.py takes [#of rolls] [#of sides on a die] and optionally [-s] which would sort the results